# campusmarket/backend/services/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CampusTrade API
  version: 1.0.0
  description: Production API contract for the CampusTrade campus marketplace.
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a campus marketplace account
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } }
      responses:
        '201': { description: Account created }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Account already exists }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate with email and password
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/google:
    post:
      tags: [Authentication]
      summary: Authenticate with a verified ICT University Google identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential: { type: string, minLength: 1, description: Google ID token }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset for an ICT University account
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UniversityEmailRequest' } } }
      responses:
        '200': { description: Generic reset-request acknowledgement }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Resend verification for an ICT University account
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UniversityEmailRequest' } } }
      responses:
        '200': { description: Generic verification-request acknowledgement }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string } } }
...
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email: { $ref: '#/components/schemas/UniversityEmail' }
        password: { type: string, minLength: 8 }
        first_name: { type: string }
        last_name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties: { email: { $ref: '#/components/schemas/UniversityEmail' }, password: { type: string } }
    UniversityEmail:
      type: string
      format: email
      maxLength: 255
      pattern: '^[a-z0-9._%+-]+@ictuniversity[.]edu[.]cm$'
      example: student@ictuniversity.edu.cm
      description: Canonical lowercase ICT University email; the API trims and lowercases accepted input.
    UniversityEmailRequest:
      type: object
      required: [email]
      properties:
        email: { $ref: '#/components/schemas/UniversityEmail' }
    AuthTokens:
      type: object
      properties: { accessToken: { type: string }, refreshToken: { type: string }, user: { type: object } }
...